% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_docid}
\alias{add_docid}
\title{Set the docid for multi-document objects}
\usage{
add_docid(x, path, docid_field)
}
\arguments{
\item{x}{data.frame; contains texts and document variables}

\item{path}{character; file path from which \code{x} is created; only use in error message}

\item{docid_field}{numeric or character; indicate position of a text column in x}

\item{impute_types}{logical; if \code{TRUE}, set types of variables automatically}
}
\description{
Set the docid for multi-document objects
}
\keyword{internal}
