% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtext-methods.R
\name{as.character.readtext}
\alias{as.character.readtext}
\title{return only the texts from a readtext object}
\usage{
\method{as.character}{readtext}(x, ...)
}
\arguments{
\item{x}{the readtext object whose texts will be extracted}

\item{...}{further arguments passed to or from other methods}
}
\description{
An accessor function to return the texts from a \link{readtext} object as a
character vector, with names matching the document names.
}
