% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_remote}
\alias{cache_remote}
\title{Internal function to cache remote file}
\usage{
cache_remote(url, ignore_missing, cache, basename = NULL,
  verbosity = 1)
}
\arguments{
\item{url}{location of a remote file}

\item{ignore_missing}{if \code{TRUE}, warns for download status}

\item{cache}{\code{TRUE}, save file in system's temporary folder and load it
from the next time}

\item{basename}{name of temporary file to preserve file extensions. If
\code{NULL}, random string will be used.}

\item{verbosity}{\itemize{
  \item 0: output errors only
  \item 1: output errors and warnings (default)
  \item 2: output a brief summary message
  \item 3: output detailed file-related messages
}}
}
\description{
Internal function to cache remote file
}
\keyword{internal}
