% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-methods.R
\name{docnames}
\alias{docnames}
\title{Extract document names from a readtext object}
\usage{
docnames(x)
}
\arguments{
\item{x}{a readtext object whose document names will be extracted}
}
\value{
a character vector of document names
}
\description{
Returns document names from a readtext object.
}
