% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-methods.R
\name{docvars}
\alias{docvars}
\alias{docvars.readtext}
\title{extract document-level metadata from a readtext object}
\usage{
docvars(x, field = NULL)

\method{docvars}{readtext}(x, field = NULL)
}
\arguments{
\item{x}{a \link{readtext} object created by \code{\link{readtxt}}}

\item{field}{string containing the document-level variable name}
}
\value{
\code{docvars} returns a data.frame of the document-level variables
}
\description{
Get the document-level metadata, also known as "docvars", from a \link{readtext} object.
}

