% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{[Experimental] Download remote file and save as a temporary file}
\usage{
download(url, cache = TRUE, verbosity = getOption("readtext_verbosity"))
}
\arguments{
\item{url}{location of a remote file}

\item{cache}{\code{TRUE}, save remote file in system's temporary folder and
load it from the next time}

\item{verbosity}{\itemize{
  \item 0: output errors only
  \item 1: output errors and warnings (default)
  \item 2: output a brief summary message
  \item 3: output detailed file-related messages
}}
}
\value{
path to a temporary file
}
\description{
[Experimental] Download remote file and save as a temporary file
}
\keyword{experimental}
\keyword{internal}
