% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_temp}
\alias{get_temp}
\title{Get path to temporary file or directory}
\usage{
get_temp(prefix = "readtext-", temp_dir = NULL, directory = FALSE,
  seed = NULL)
}
\arguments{
\item{prefix}{a string appended to random file or directory names.}

\item{temp_dir}{a path to temporary directory. If \code{NULL}, value from \code{tempdir()} will be used.}

\item{directory}{If \code{TRUE}, temporary directory wil be created.}

\item{seed}{a seed value for \code{digest::digest}. If code{NULL}, a random value will be used.}
}
\description{
Get path to temporary file or directory
}
\keyword{internal}
