% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtext-methods.R
\name{print.readtext}
\alias{print.readtext}
\title{print method for a readtext object}
\usage{
\method{print}{readtext}(x, n = 6L, text_width = 10L, ...)
}
\arguments{
\item{x}{the readtext object to be printed}

\item{n}{a single integer, the number of rows of a readtext object to print.}

\item{text_width}{number of characters to display of the text field}

\item{...}{not used here}
}
\description{
Print a readtext object in a nicely formatted way.
}
