% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_fields}
\alias{sort_fields}
\title{Move text to the first column and set types to document variables}
\usage{
sort_fields(x, path, text_field, impute_types = TRUE)
}
\arguments{
\item{x}{data.frame; contains texts and document variables}

\item{paht}{character; file path from which \code{x} is created; only use in error message}

\item{text_filed}{numeric or character; indicate position of a text column in x}

\item{inpute_types}{logical; if \code{TRUE}, set types of variables automatically}
}
\description{
Move text to the first column and set types to document variables
}
\keyword{internal}
