% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-methods.R
\name{texts}
\alias{texts}
\alias{texts.readtext}
\title{extract texts from a readtext object}
\usage{
texts(x, groups = NULL, ...)

\method{texts}{readtext}(x, groups = NULL, ...)
}
\arguments{
\item{x}{a \link{readtext} object created by \code{\link{readtext}}}

\item{groups}{character vector containing the names of document variables in 
a corpus, or a factor equal in length to the number of documents, used for 
aggregating the texts through concatenation.  If \code{x} is of type
character, then \code{groups} must be a factor.}

\item{...}{unused}
}
\value{
For \code{texts}, a character vector of the texts in the corpus.
  
  For \code{texts <-}, the corpus with the updated texts.
}
\description{
Get the texts in a text or corpus object, with grouping options. Works for
plain character vectors too, if \code{groups} is a factor.
}

