% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-methods.R
\name{docvars}
\alias{docvars}
\title{Extract document variables from a readtext object}
\usage{
docvars(x)
}
\arguments{
\item{x}{a \link{readtext} object whose document variables will be extracted}
}
\value{
a data.frame of document variables
}
\description{
Returns document variables from a readtext object.
}
