% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-methods.R
\name{texts}
\alias{texts}
\alias{texts.readtext}
\title{Get corpus texts [deprecated]}
\usage{
texts(x, ...)

\method{texts}{readtext}(x, ...)
}
\arguments{
\item{x}{a \link{readtext} object}

\item{...}{not used}

\item{spacer}{when concatenating texts by using \code{groups}, this will be the
spacing added between texts.  (Default is two spaces.)}
}
\value{
a character vector of the texts in the corpus
}
\description{
Get the texts from a \link{readtext} object.
}
\details{
This function is \strong{deprecated}.

Use \code{\link[=as.character.readtext]{as.character.readtext()}} to turn a readtext object into a simple named
character vector of documents.
}
\keyword{corpus}
\keyword{internal}
